Split rhythm classes into individual files
