+++ wheatley/rhythm/__init__.py
@@ -1,0 +1,5 @@
+""" A module to contain all the different ways a 'rhythm' can be derived whilst Wheatley is ringing. """
+
+from .abstract_rhythm import Rhythm
+from .regression import RegressionRhythm
+from .wait_for_user import WaitForUserRhythm
